requirements.txt
numpy==1.24.3
pandas==2.0.2
scipy==1.10.1
matplotlib==3.7.1
seaborn==0.12.2
setuptools==67.8.0
plotly==5.15.0
kaleido==0.2.1
argparse==1.4.0
numba~=0.57.0
PyYAML~=6.0
scikit-learn==1.2.2
statsmodels==0.14.0
ruptures~=1.1.7
IPython~=8.14.0
orca~=1.8
dash~=2.13.0
pymongo~=4.6.1
pika~=1.3.2
fastdtw~=0.3.4
dtaidistance~=2.3.10